Fix for givelum command

+++ Source/ClientCommands.cpp
@@ -4989,7 +4989,7 @@ CLIENT_COMMAND(givelum, "[value]", "Gives you some Luminance for testing.", BASI
 		amount = strtoul(argv[0], NULL, 10);
 	}
 
-	amount = min(amount, 1000000);
+	amount = min(amount, (DWORD)1000000);
 	pPlayer->m_Qualities.SetInt64(AVAILABLE_LUMINANCE_INT64, amount + total);
 	pPlayer->NotifyInt64StatUpdated(AVAILABLE_LUMINANCE_INT64, false);
